Split some initial modules files to use cython on them.
